Added checking condition for type of given line instances
What changes you have made?
Using instanceof while checking that lines are equal or not.
And added test cases according to it.

Why this change is necessary?
It will not give for the lines are equal or not true
when the given lines are not of the same instance.

What is the impact of the change?
Deep checking.